+++ src/fixIcelandicLocale.ts
@@ -1,8 +1,10 @@
 import {
   _PatchedCollator,
   _PatchedDateTimeFormat,
+  _PatchedListFormat,
   _patchedLocaleCompare,
   _PatchedNumberFormat,
+  _PatchedPluralRules,
   _patchedToLocaleDateString,
   _patchedToLocaleString,
 } from './fixIcelandicLocale.privates.js';
@@ -23,3 +25,11 @@ if (Intl.Collator.supportedLocalesOf(['is']).length < 1) {
   Intl.DateTimeFormat = _PatchedDateTimeFormat;
   Date.prototype.toLocaleDateString = _patchedToLocaleDateString;
 }
+/* eslint-disable @typescript-eslint/no-unnecessary-type-assertion, @typescript-eslint/no-unnecessary-condition */
+if (Intl.ListFormat && Intl.ListFormat.supportedLocalesOf(['is']).length < 1) {
+  (Intl.ListFormat as typeof Intl.ListFormat) = _PatchedListFormat;
+}
+if (Intl.PluralRules && Intl.PluralRules.supportedLocalesOf(['is']).length < 1) {
+  (Intl.PluralRules as typeof Intl.PluralRules) = _PatchedPluralRules;
+}
+/* eslint-enable @typescript-eslint/no-unnecessary-type-assertion, @typescript-eslint/no-unnecessary-condition */