+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `@reykjavik/webtools/fixIcelandicLocale`:
+  - feat: Patch `Intl.PluralRules` and `Intl.ListFormat`
 
 ## 0.1.16 – 0.1.17